chore: Extract scope manager single continuation
